requirements.txt
python-dotenv==1.0.0
gspread==4.0.1
oauth2client==4.1.3
selenium==4.13.0
